Apply(feature) : create functions for input

+++ app/Views/admin/setting.php
@@ -4,7 +4,7 @@
             Setting
         </h3>
         <div class="control-wrap">
-            <a href="javascript:create();" class="button create">
+            <a href="javascript:openPopupCreate();" class="button create">
                 <img src="/asset/images/icon/plus.png"/>
                 <span>Create</span>
             </a>
@@ -72,21 +72,26 @@
                     html += extracted;
                 }
             }
-            html += `
-            <div class="control-wrap line-before">
-                <a href="javascript:edit(${data['id']})"
-                   class="button edit">
-                    <img src="/asset/images/icon/edit.png"/>
-                    <span>Edit</span>
-                </a>
-                <a href="javascript:openPopupDelete(${data['id']});" class="button delete">
-                    <img src="/asset/images/icon/delete.png"/>
-                    <span>Delete</span>
-                </a>
-            </div>`;
 
             html += `</div>`;
             return html;
         },
+        getControlHtml: function (data) {
+            let html = `<div class="control-wrap line-before">
+                <a href="javascript:edit(${data['id']})"
+                   class="button edit">
+                    <img src="/asset/images/icon/edit.png"/>
+                    <span>Edit</span>
+                </a>`;
+            if (data['is_deletable'] == 1) {
+                html += `
+                <a href="javascript:openPopupDelete(${data['id']});" class="button delete">
+                    <img src="/asset/images/icon/delete.png"/>
+                    <span>Delete</span>
+                </a>`;
+            }
+            html += `</div>`;
+            return html;
+        }
     })
 </script>
